Enable templating within a famous:html-element
Allow users to use Mustache templates within their famous:html-elements.
This adds a couple of pieces to the state bag. Locals is an object of
the variables we want to interpolate into the template. initialContent is
the initial state of the content so we can re-template against it when
the locals change. And didTemplate is a boolean to check whether we have
already rendered or not.